+++ client/src/components/Footer.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import './footer.css';
+import '../index.css';
 
 export default function Footer() {
   return (
