+++ client/src/components/HomeCard.jsx
@@ -6,7 +6,7 @@ import CardMedia from "@mui/material/CardMedia";
 import Button from "@mui/material/Button";
 import Typography from "@mui/material/Typography";
 import Rating from "@mui/material/Rating";
-import "./homeCard.css";
+import "../index.css";
 import Icon from "@mui/material/Icon";
 import CheckIcon from "@mui/icons-material/Check";
 import ClearIcon from "@mui/icons-material/Clear";
@@ -32,7 +32,7 @@ export default function HomeCard({ landlord }) {
               style={{ fontFamily: "Nunito" }}
               color="#fff"
             >
-              {landlord.full_name}
+              {landlord.first_name} {landlord.last_name}
             </Typography>
             <Typography
               variant="span"
