--- client/src/components/clearNavbar/clearNavbar.css
@@ -1,28 +1,0 @@
-#clearNavBar {
-  display: flex;
-  justify-content: space-between;
-  background-color: none;
-  padding: 15px;
-  position: absolute;
-  width: 100%;
-}
-
-#logo {
-  font-family: "Nunito", sans-serif;
-  font-size: xx-large;
-  font-style: bold;
-  color: tomato;
-  font-weight: 700; 
-}
-
-.clearNavBarListItems {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.clearNavBarListItem {
-  list-style: none;
-  margin-left: 50px;
-  color: #666;
-}
