--- client/src/components/footer/footer.css
@@ -1,9 +1,0 @@
-#footer {
-  height: 40px;
-  background-color: #333;
-  color: rgba(255, 255, 255, 0.514);
-  font-size: 0.75em;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
