--- client/src/components/homeCard/homeCard.css
@@ -1,18 +1,0 @@
-.homeCardContents {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.homeCardRating {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 10px;
-}
-
-.bikePetFriendlyRating {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
