--- client/src/components/homeCards/homeCards.css
@@ -1,38 +1,0 @@
-.homeCardsWrapper {
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.homeCards {
-  display: flex;
-  justify-content: center;
-  align-items: flex-start;
-  flex-wrap: wrap;
-  margin-bottom: 25px;
-}
-
-.homeCardsUserFeedbackWrapper,
-.homeCardsUserFeedback {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.homeCardsUserFeedback {
-  margin-bottom: 25px;
-}
-
-.feedbackDesc,
-.feedbackUser {
-  color: white;
-  font-family: "Nunito", sans-serif;
-}
-
-.feedbackDesc {
-  font-style: italic;
-  font-size: 1.1em;
-}
