--- client/src/components/login/login.css
@@ -1,8 +1,0 @@
-.linkButton {
-  color: blueviolet;
-  text-decoration: underline;
-}
-
-.linkButton:hover {
-  color: blue;
-}
