+++ client/src/components/tomatopalette.jsx
@@ -1,0 +1,13 @@
+import { createTheme, ThemeProvider } from '@mui/material/styles';
+
+// create our color palette
+const tomatopalette = createTheme({
+  palette: {
+    primary: {
+      main: '#ff6347',
+      contrastText: '#fff',
+    },
+  },
+});
+
+export default tomatopalette;
