+++ client/src/index.css
@@ -1,0 +1,313 @@
+#clearNavBar {
+  display: flex;
+  justify-content: space-between;
+  background-color: none;
+  padding: 15px;
+  position: absolute;
+  width: 100%;
+}
+
+#logo {
+  font-family: "Nunito", sans-serif;
+  font-size: xx-large;
+  font-style: bold;
+  color: tomato;
+  font-weight: 700; 
+}
+
+.clearNavBarListItems {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.clearNavBarListItem {
+  list-style: none;
+  margin-left: 50px;
+  color: #666;
+}
+
+#footer {
+  height: 40px;
+  background-color: #333;
+  color: rgba(255, 255, 255, 0.514);
+  font-size: 0.75em;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.homeCardContents {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.homeCardRating {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 10px;
+}
+
+.bikePetFriendlyRating {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.homeCardsWrapper {
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.homeCards {
+  display: flex;
+  justify-content: center;
+  align-items: flex-start;
+  flex-wrap: wrap;
+  margin-bottom: 25px;
+}
+
+.homeCardsUserFeedbackWrapper,
+.homeCardsUserFeedback {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.homeCardsUserFeedback {
+  margin-bottom: 25px;
+}
+
+.feedbackDesc,
+.feedbackUser {
+  color: white;
+  font-family: "Nunito", sans-serif;
+}
+
+.feedbackDesc {
+  font-style: italic;
+  font-size: 1.1em;
+}
+
+.linkButton {
+  color: blueviolet;
+  text-decoration: underline;
+}
+
+.linkButton:hover {
+  color: blue;
+}
+
+#navBar {
+  height: 50px;
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  background-color: lightgrey;
+  /* background: none; */
+  align-items: center;
+  /* position: absolute; */
+}
+
+.navBarLeft {
+  flex: 1;
+  display: flex;
+  /* justify-content: center; */
+  padding-left: 25px;
+  align-items: center;
+}
+
+.navBarCenter {
+  flex: 3;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.navBarListItems {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-size: 1.3em;
+}
+
+.navBarListItem {
+  list-style: none;
+  font-family: 'Nunito', sans-serif;
+  padding: 0 20px;
+}
+
+.navBarRight {
+  flex: 1;
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  padding-right: 25px;
+} */
+
+
+a:hover {
+  color: tomato;
+} 
+a {
+  font-family: 'Nunito';
+  text-decoration: none;
+  padding-right: 10px;
+  list-style: none;
+  color: #666;
+}
+
+h3 {
+  margin-bottom: 15px;
+}
+.css-binzgt {
+  margin-top:  0px;
+}
+.exitButton {
+  margin-bottom: 10px;
+  background-color: tomato;
+  padding: 5px 10px 5px 10px;
+  border: none;
+  border-radius: 3px;
+}
+.exitButton:hover {
+  color: white;
+}
+
+.linkButton{
+  margin-bottom:  10px;
+}
+
+#loginSignup {
+  position: absolute;
+  /* top: 50%;
+  left: 50%; */
+  /* transform: translate(-50%, -50%); */
+  background-color: white;
+  /* border: 2px black; */
+  border-radius: 5px;
+  padding: 25px;
+  z-index: 2;
+}
+
+.home {
+  min-height: 100vh;
+  background-image: url("http://localhost:3000/images/marten-bjork-4WBIyO_2SbU-unsplash.jpg");
+  background-repeat: no-repeat;
+  background-size: cover;
+}
+
+.homeBanner {
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.homeTitle {
+  font-family: "Nunito", sans-serif;
+  color: white;
+  text-align: center;
+  font-size: 20px;
+}
+
+.homeTitleText {
+  color: tomato;
+  font-size: 50px;
+}
+
+.homeTitleIcon {
+  color: tomato;
+}
+
+#map { 
+  height: 90vh;
+  margin-left: 100px;
+  margin-right: 100px;
+ }
+
+.leaflet-container {
+  height: 100%;
+  z-index: 0;
+
+}
+
+#background {
+  min-height: 100vh;
+  width: auto;
+  background-image: url("http://localhost:3000/images/marten-bjork-4WBIyO_2SbU-unsplash.jpg");
+  background-repeat: no-repeat;
+  background-size: cover;
+  background-attachment: fixed;
+}
+
+.reviewPageGlobalContainer {
+  margin: 0;
+  min-height: 100vh;
+  background-image: url("http://localhost:3000/images/marten-bjork-4WBIyO_2SbU-unsplash.jpg");
+  background-repeat: no-repeat;
+  background-size: cover;
+}
+
+.reviewformContainer{
+  background: rgba(241, 241, 241, 0.6);
+}
+
+.reviewMainContainer {
+  text-decoration: none;
+  font-family: "Roboto","Helvetica","Arial",sans-serif;
+  color: rgb(68, 67, 67);
+}
+
+.reviewLabel{
+  margin: 0;
+  text-align: end;
+  justify-content: center;
+  align-items: center;
+  padding-top:4px
+}
+
+.searchMainContainer {
+  text-decoration: none;
+  font-family: "Roboto","Helvetica","Arial",sans-serif;
+  color: rgb(68, 67, 67);
+}
+
+.searchContainer{
+  background: rgba(241, 241, 241, 0.4)
+}
+.searchPageMain {
+  margin: 0;
+  min-height: 100vh;
+  background-image: url("http://localhost:3000/images/marten-bjork-4WBIyO_2SbU-unsplash.jpg");
+  background-repeat: no-repeat;
+  background-size: cover;
+  background-attachment: fixed;
+}
+
+#userProfile {
+  padding: 25px 100px 25px 100px;
+  min-height: 100vh;
+  background-image: url("http://localhost:3000/images/marten-bjork-4WBIyO_2SbU-unsplash.jpg");
+  background-repeat: no-repeat;
+  background-size: cover;
+  background-attachment: fixed;
+  
+}
+
+#userProfileTitle {
+
+  font-family: 'Nunito';
+  color: tomato;
+  margin-bottom: 50px;
+}
+h4 {
+  margin-bottom: 20px;
+}
+
