--- client/src/pages/profile/profile.css
@@ -1,8 +1,0 @@
-#background {
-  min-height: 100vh;
-  width: auto;
-  background-image: url("http://localhost:3000/images/marten-bjork-4WBIyO_2SbU-unsplash.jpg");
-  background-repeat: no-repeat;
-  background-size: cover;
-  background-attachment: fixed;
-}
