--- client/src/pages/search/search.css
@@ -1,17 +1,0 @@
-.searchMainContainer {
-    text-decoration: none;
-    font-family: "Roboto","Helvetica","Arial",sans-serif;
-    color: rgb(68, 67, 67);
-}
-
-.searchContainer{
-    background: rgba(241, 241, 241, 0.4)
-}
-.searchPageMain {
-    margin: 0;
-    min-height: 100vh;
-    background-image: url("http://localhost:3000/images/marten-bjork-4WBIyO_2SbU-unsplash.jpg");
-    background-repeat: no-repeat;
-    background-size: cover;
-    background-attachment: fixed;
-}
