--- client/src/pages/user/userprofile.css
@@ -1,19 +1,0 @@
-#userProfile {
-  padding: 25px 100px 25px 100px;
-  min-height: 100vh;
-  background-image: url("http://localhost:3000/images/marten-bjork-4WBIyO_2SbU-unsplash.jpg");
-  background-repeat: no-repeat;
-  background-size: cover;
-  background-attachment: fixed;
-  
-}
-
-#userProfileTitle {
-
-  font-family: 'Nunito';
-  color: tomato;
-  margin-bottom: 50px;
-}
-h4 {
-  margin-bottom: 20px;
-}
