+++ package.json
@@ -7,7 +7,8 @@
     "test": "",
     "start": "nodemon server/server.js",
     "build": " NODE_ENV=production webpack ",
-    "dev": "concurrently 'cross-env NODE_ENV=development webpack serve --open' 'nodemon server/server.js'"
+    "dev": "concurrently \"cross-env NODE_ENV=development webpack serve --open\" \"nodemon server/server.js\""
+
   },
   "repository": {
     "type": "git",
